refactor: relocate chartUtils to lib/ for architectural consistency
